size of type

+++ 0x00-hello_world/6-size.c
@@ -1,11 +1,8 @@
 #include <stdio.h>
-
 /**
  * main - the main function
- *
  * Return: 0 when successful
  */
-
 int main(void)
 {
 	char a;
@@ -17,7 +14,7 @@ int main(void)
 printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(a));
 printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(b));
 printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(c));
-printf("Size of a long long int: %lu byte(s)\n", (unsigned long long)sizeof(d));
+printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(d));
 printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
 return (0);
 }
